Move mobile tabs behaviour to separate file
This allows us to simplify header.js and hopefully makes the code easier to maintain
